pyproject.toml
[project]
name = "youtube-comments"
version = "0.1.0"
description = ""
authors = [
    {name = "andreas.arayaosorio",email = "andreas.arayaosorio@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-api-python-client (>=2.160.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
